+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Changes since v7.15.4
 
+- [#2998](https://github.com/oauth2-proxy/oauth2-proxy/issues/2998) feat: add support for partitioned cookies
+
 # V7.15.4
 
 ## Release Highlights